Fixed documentation
