+++ src/scenario.py
@@ -1,20 +1,84 @@
+import subprocess
+
 import psycopg2
 from tqdm import tqdm
 
+from config import DDLStep
 from db.yugabyte import ENABLE_STATISTICS_HINT
 from models.factory import get_test_model
-from utils import evaluate_sql, calculate_avg_execution_time, \
-    get_md5
+from utils import evaluate_sql, calculate_avg_execution_time, get_md5
 
 
-class QueryEvaluator:
+class Scenario:
     def __init__(self, config):
         self.config = config
         self.logger = self.config.logger
+        self.sut_database = self.config.database
 
-    def evaluate(self,
-                 connection,
-                 evaluate_optimizations=False):
+    def start_db(self):
+        self.logger.info(f"Initializing {self.sut_database.__class__.__name__} DB")
+
+        commit_hash = self.config.revision
+
+        self.sut_database.change_version_and_compile(commit_hash)
+        self.sut_database.stop_database()
+        self.sut_database.destroy()
+        self.sut_database.start_database()
+
+        return self.get_commit_message(commit_hash)
+
+    def get_commit_message(self, commit_hash):
+        if commit_hash:
+            output = str(subprocess.check_output(
+                f"echo `git log -n 1 --pretty=format:%s {commit_hash}`",
+                cwd=self.config.source_path,
+                shell=True)).rstrip('\n')
+            return f"{output} ({commit_hash})"
+        else:
+            return ""
+
+    def stop_db(self):
+        self.sut_database.stop_database()
+
+    def evaluate(self):
+        loader = self.config.database.get_results_loader()
+
+        commit_message = self.start_db()
+        try:
+            test_database = self.config.connection.database
+            self.create_test_database(test_database)
+
+            self.sut_database.establish_connection(test_database)
+
+            loq = self.config.database.get_list_queries()
+            loq.db_version = self.sut_database.connection.get_version()
+            loq.model_queries, loq.queries = self.run_ddl_and_testing_queries(
+                self.sut_database.connection.conn, self.config.with_optimizations)
+            loq.git_message = commit_message
+
+            self.logger.info(f"Storing results to report/{self.config.output}")
+            loader.store_queries_to_file(loq, self.config.output)
+        except Exception as e:
+            self.logger.exception(e)
+            raise e
+        finally:
+            if self.config.clean_db:
+                self.stop_db()
+
+    def create_test_database(self, test_database):
+        if DDLStep.DATABASE in self.config.ddls:
+            self.sut_database.establish_connection("postgres")
+            conn = self.sut_database.connection.conn
+            try:
+                with conn.cursor() as cur:
+                    colocated = "" if self.config.ddl_prefix else " WITH COLOCATED = true"
+                    evaluate_sql(cur, f'CREATE DATABASE {test_database}{colocated};')
+            except Exception as e:
+                self.logger.exception(f"Failed to create testing database {e}")
+
+    def run_ddl_and_testing_queries(self,
+                                    connection,
+                                    evaluate_optimizations=False):
         queries = []
         model_queries = []
         try:
@@ -43,7 +107,8 @@ class QueryEvaluator:
                     evaluate_sql(cur, ENABLE_STATISTICS_HINT)
 
                 try:
-                    evaluate_sql(cur, f"SET statement_timeout = '{self.config.test_query_timeout}s'")
+                    evaluate_sql(cur,
+                                 f"SET statement_timeout = '{self.config.test_query_timeout}s'")
 
                     short_query = original_query.query.replace('\n', '')[:40]
                     self.logger.info(
@@ -51,20 +116,23 @@ class QueryEvaluator:
 
                     try:
                         evaluate_sql(cur, original_query.get_explain())
-                        original_query.execution_plan = self.config.database.get_execution_plan('\n'.join(
-                            str(item[0]) for item in cur.fetchall()))
+                        original_query.execution_plan = self.config.database.get_execution_plan(
+                            '\n'.join(
+                                str(item[0]) for item in cur.fetchall()))
 
                         conn.rollback()
                     except psycopg2.errors.QueryCanceled:
                         try:
                             evaluate_sql(cur, original_query.get_heuristic_explain())
-                            original_query.execution_plan = self.config.database.get_execution_plan('\n'.join(
-                                str(item[0]) for item in cur.fetchall()))
+                            original_query.execution_plan = self.config.database.get_execution_plan(
+                                '\n'.join(
+                                    str(item[0]) for item in cur.fetchall()))
 
                             conn.rollback()
                         except psycopg2.errors.QueryCanceled:
                             self.logger.error("Unable to get execution plan even w/o analyze")
-                            original_query.execution_plan = self.config.database.get_execution_plan('')
+                            original_query.execution_plan = self.config.database.get_execution_plan(
+                                '')
 
                     calculate_avg_execution_time(cur, original_query,
                                                  num_retries=int(self.config.num_retries),
@@ -93,7 +161,8 @@ class QueryEvaluator:
         self.logger.debug(f"{len(list_of_optimizations)} optimizations generated")
         progress_bar = tqdm(list_of_optimizations)
         num_skipped = 0
-        min_execution_time = original_query.execution_time_ms if original_query.execution_time_ms > 0 else (self.config.test_query_timeout * 1000)
+        min_execution_time = original_query.execution_time_ms if original_query.execution_time_ms > 0 else (
+                    self.config.test_query_timeout * 1000)
         original_query.optimizations = []
         execution_plans_checked = set()
